combine IssueToken.vue and BurnToken.vue to one component IssueAndBurnToken.vue
